+++ AppSettings/ApplicationSettings.cs
@@ -108,8 +108,8 @@ namespace PlenBotLogUploader.AppSettings
         [JsonProperty("upload")]
         internal ApplicationSettingsUpload Upload { get; set; } = new ApplicationSettingsUpload();
 
-		[JsonProperty("buildCodes")]
-		internal ApplicationSettingsBuildCodes BuildCodes { get; set; } = new();
-		#endregion
-	}
+        [JsonProperty("buildCodes")]
+        internal ApplicationSettingsBuildCodes BuildCodes { get; set; } = new();
+        #endregion
+    }
 }
