+++ AppSettings/ApplicationSettingsBuildCodes.cs
@@ -5,8 +5,8 @@ namespace PlenBotLogUploader.AppSettings
 {
     internal sealed class ApplicationSettingsBuildCodes
     {
-		[JsonProperty("demoteRunes")]  internal bool               DemoteRunes  { get; set; } = true;
-		[JsonProperty("demoteSigils")] internal bool               DemoteSigils { get; set; } = true;
-		[JsonProperty("compression")]  internal CompressionOptions Compression  { get; set; } = CompressionOptions.ALL;
+        [JsonProperty("demoteRunes")] internal bool DemoteRunes { get; set; } = true;
+        [JsonProperty("demoteSigils")] internal bool DemoteSigils { get; set; } = true;
+        [JsonProperty("compression")] internal CompressionOptions Compression { get; set; } = CompressionOptions.ALL;
     }
 }
